Factor out into testingColumnTexts
