+++ app/assets/javascripts/student_profile/LightProfilePage.js
@@ -19,9 +19,8 @@ import LightNotesHelpContext from './LightNotesHelpContext';
 import StudentSectionsRoster from './StudentSectionsRoster';
 import {tags} from './lightTagger';
 import DetailsSection from './DetailsSection';
-import {toMoment} from './QuadConverter';
-import {shortLabelFromNextGenMcasScore, shortLabelFromOldMcasScore} from './mcasScores';
 import FullCaseHistory from './FullCaseHistory';
+import testingColumnTexts from './testingColumnTexts';
 
 
 // Prototype of profile v3
@@ -526,49 +525,4 @@ function percentileWithSuffix(percentile) {
     3: 'rd'
   }[lastDigit] || 'th';
   return `${percentile}${suffix}`;
-}
-
-
-// Look at ELA and Math next gen MCAS scores and make the different texts for the 'testing'
-// column summary for HS.  If those don't exist, fall back to old-school MCAS.
-function testingColumnTexts(nowMoment, chartData) {
-  const ela = (
-    latestMcasScoreSummary(chartData.next_gen_mcas_ela_scaled, shortLabelFromNextGenMcasScore, nowMoment) ||
-    latestMcasScoreSummary(chartData.mcas_series_ela_scaled, shortLabelFromOldMcasScore, nowMoment) ||
-    missingMcasSummary()
-  );
-  const math = (
-    latestMcasScoreSummary(chartData.next_gen_mcas_mathematics_scaled, shortLabelFromNextGenMcasScore, nowMoment) ||
-    latestMcasScoreSummary(chartData.mcas_series_math_scaled, shortLabelFromOldMcasScore, nowMoment) ||
-    missingMcasSummary()
-  );
-
-  const scoreText = ela.scoreText === math.scoreText
-    ? ela.scoreText
-    : `${ela.scoreText} / ${math.scoreText}`;
-  const testText = ela.scoreText === math.scoreText
-    ? 'ELA and Math MCAS'
-    : 'ELA / Math MCAS';
-  const dateText = (ela.dateText === math.dateText)
-    ? ela.dateText
-    : `${ela.dateText} / ${math.dateText}`;
-
-  return {scoreText, testText, dateText};
-}
-
-// Make text for the score and date for the latest MCAS score (either next gen
-// or old, depending on `scoreTextFn`).
-function latestMcasScoreSummary(quads, scoreTextFn, nowMoment) {
-  const latestEla = _.last(_.sortBy(quads || [], quad => toMoment(quad).unix()));
-  if (!latestEla) return null;
-
-  const scoreText = scoreTextFn(latestEla[3]);
-  const dateText = toMoment(latestEla).from(nowMoment);
-  return {scoreText, dateText};
-}
-
-function missingMcasSummary() {
-  const scoreText = '-';
-  const dateText = 'not yet taken';
-  return {scoreText, dateText};
 }
